requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.17.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.0
Babel==2.14.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.3
bleach==6.1.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
Brotli==1.1.0
browser-use==0.4.4
bs4==0.0.2
bubus==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.2
Crawl4AI==0.6.2
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==4.4.2
deepseek==1.0.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
Django==5.1.6
django-admin-thumbnails==0.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django_paystack==1.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
doc2text==0.2.4
docx2txt==0.8
easyocr==1.7.2
ecdsa==0.19.0
editdistance==0.8.1
efficientnet==1.0.0
email_validator==2.2.0
essential-generators==1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.11.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.115.12
fastjsonschema==2.19.1
filelock==3.14.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.60.1
fpdf2==2.8.5
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.80.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
google-genai==1.21.1
google-generativeai==0.8.5
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.69.2
googlesearch-python==1.3.0
gotrue==2.4.4
greenlet==3.2.4
griffe==1.7.3
groq==0.29.0
grpcio==1.72.1
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.3
hyperframe==6.1.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
IMAPClient==3.0.1
imgaug==0.4.0
importlib_metadata==8.6.1
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.22
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.10.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.45
lazy_loader==0.4
Levenshtein==0.27.1
lib-py-parse==0.0.1
libclang==18.1.1
litellm==1.67.5
llama-cloud==0.1.34
llama-cloud-services==0.6.51
llama-index-core==0.12.51
llama-index-instrumentation==0.3.0
llama-index-workflows==1.1.0
llama-parse==0.6.51
lvldb==0.1.2
lxml==5.3.1
lxml_html_clean==0.4.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mem0ai==0.1.93
mime==0.1.0
mistune==3.0.2
ml_dtypes==0.5.1
monotonic==1.6
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msal==1.34.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.33.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
notebook==7.1.1
notebook_shim==0.2.4
numpy==2.3.5
o365==2.1.7
oauthlib==3.2.2
ollama==0.5.1
openai==2.8.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.3
passlib==1.7.4
patchright==1.52.5
paystackapi==2.1.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.3.8
playwright==1.52.0
plotly==6.1.2
portalocker==2.10.1
postgrest==0.16.8
posthog==3.25.0
proglog==0.1.12
prometheus_client==0.20.0
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.21.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.26.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypaystack2==2.0.3
PyPDF2==3.0.1
pypdfium2==5.1.0
pyperclip==1.9.0
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyquery==2.0.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.3.0
python-json-logger==2.0.7
python-Levenshtein==0.27.1
python-multipart==0.0.18
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.14.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
realtime==1.0.5
redis==7.0.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.5
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
s3transfer==0.13.0
scikit-image==0.25.2
scipy==1.14.1
screeninfo==0.8.1
selenium==4.30.0
Send2Trash==1.8.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
storage3==0.7.6
streamlit==1.44.1
streamlit-aggrid==1.1.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.6
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.0
tf-playwright-stealth==1.1.2
tifffile==2025.6.11
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.1
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.8.19.20240311
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.158
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.34.3
validators==0.35.0
vine==5.1.0
virtualenv==20.26.2
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.14
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
